Move indexing impls into kernels::indexing
